[chore]: Replace toast with snackbar

+++ app/src/main/java/com/mythio/weather/model/entity/ForecastW.kt
@@ -15,7 +15,7 @@ data class ForecastW(
     var date: String,
 
     @Json(name = "date_epoch")
-    var dateEpoch: Int,
+    var dateEpoch: Long,
 
     @Json(name = "day")
     @Embedded(prefix = "day_")
